refactor: Unify CreateTestFile and CreateTestFileWithContent
